move parseJson out to a shared helper
